Fix OF-Config v1.1 XML Schema by changing controller's ip-address type
There is a discrepancy in the XML Schema and YANG model in
OFControllerType's child element in OF-Config v1.1. The former one
defines type of ip-address element as inet:ip-prefix while the latter
one claims that the type is inet:ip-address. This commit fixes
this ambiguity.

It also updates the test appropriately.

(Motivated by FlowForwarding/LINC-Switch#305).